% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{changepoint_CI}
\alias{changepoint_CI}
\title{Confidence Interval}
\usage{
changepoint_CI(
  decomp,
  alpha = 0.05,
  B = 5000,
  K = 500,
  regime = c("auto", "vanishing", "nonvanishing")
)
}
\description{
Confidence Interval
}
\details{
to be added
}
