% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{decompose_signal}
\alias{decompose_signal}
\title{Signal Decomposition}
\usage{
decompose_signal(
  y,
  diff,
  w = 200,
  ma_window = 100,
  right_tail_cutoff = 0.95,
  left_tail_cutoff = 0.6,
  threshold = "auto"
)
}
\description{
Signal Decomposition
}
\details{
to be added
}
