% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{fit_mlp}
\alias{fit_mlp}
\title{Fits MLP with defined structure to rolling window of defined size,
Then fits another MLP with double the rolling window,
capturing two windows' uncertainty}
\usage{
fit_mlp(
  vec,
  w = 100,
  act1 = "Act_TanH_Xdiv2",
  act2 = "Act_TanH_Xdiv2",
  lr1 = 0.01,
  lr2 = 0.01,
  hl1 = 16,
  hl2 = 32,
  ep1 = 1000,
  ep2 = 2000
)
}
\arguments{
\item{vec}{original data}

\item{act1}{activation function of smaller MLP}

\item{act2}{activation function of larger MLP}

\item{lr1, lr2}{learning rates of smaller and larger MLP}

\item{hl1, hl2}{number of neurons in the hidden layers of smaller and larger MLP}

\item{ep1, ep2}{number of training epochs for smaller and larger MLP}
}
\value{
a list of two elements containing the fitted values for the smaller
and larger MLP models, original input and index.
}
\description{
Fits MLP with defined structure to rolling window of defined size,
Then fits another MLP with double the rolling window,
capturing two windows' uncertainty
}
