% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{ma}
\alias{ma}
\title{Two-sided moving average smoothing}
\usage{
ma(x, n = 5, circular = T)
}
\arguments{
\item{x}{a numeric vector}

\item{n}{number of neighbors to smooth over in total}

\item{circular}{if the ends should wrap or remain NA}
}
\value{
smoothed numeric vector
}
\description{
Two-sided moving average smoothing
}
