% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_mlp.R
\name{plot_mlp_slider_3lines}
\alias{plot_mlp_slider_3lines}
\title{Interactive plot of fit_mlp output}
\usage{
plot_mlp_slider_3lines(
  y,
  fit_mlp_res,
  t.chp.ind = NA,
  step = 50,
  start = 75,
  w = 100
)
}
\arguments{
\item{y}{the original vector of observations}

\item{fit_mlp_res}{the resulting list of fit_mlp function}
}
\value{
plotly object
}
\description{
Interactive plot of fit_mlp output
}
