% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{select_best_spike}
\alias{select_best_spike}
\title{Best Spike Select}
\usage{
select_best_spike(s, right_tail_cutoff = 0.95, left_tail_cutoff = 0.6)
}
\description{
Best Spike Select
}
\details{
to be added
}
