% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{calc_detector}
\alias{calc_detector}
\title{Calculate Detector Statistic}
\usage{
calc_detector(y, fit_mlp_res, w = 100, a = 1, b = 0)
}
\arguments{
\item{y}{original data}

\item{fit_mlp_res}{output of fit_mlp function}

\item{w}{window size}

\item{a}{a numeric value between 0 and 1, 0 indicating only ratio, 1 indicating only difference}

\item{b}{correction to discontinuity for ratio}
}
\value{
a vector of detector statistic
}
\description{
Calculate Detector Statistic
}
\details{
Calculates detector statistic using a linear combination of Ratio and DIfference
between RSS of full MLP and half window MLP
}
